+++ custom_components/opendisplay_studio/widgets/entity_state/__init__.py
@@ -1,1 +1,0 @@
-"""Entity State widget package."""