+++ custom_components/opendisplay_studio/widgets/sensor/__init__.py
@@ -1,0 +1,1 @@
+"""Sensor widget package."""